Slang Migration Finish-Up

+++ ashura/engine/gpu_system.h
@@ -11,7 +11,7 @@ namespace ash
 {
 
 /// @details do not change the underlying type. It maps directly to the GPU handle
-enum class TextureId : u16
+enum class TextureId : u32
 {
   Base        = 0,
   White       = 0,
@@ -26,10 +26,10 @@ enum class TextureId : u16
   Yellow      = 9
 };
 
-inline constexpr u16 NUM_DEFAULT_TEXTURES = 10;
+inline constexpr u32 NUM_DEFAULT_TEXTURES = 10;
 
 /// @details do not change the underlying type. It maps directly to the GPU handle
-enum class SamplerId : u16
+enum class SamplerId : u32
 {
   Linear         = 0,
   Nearest        = 1,
@@ -37,7 +37,7 @@ enum class SamplerId : u16
   NearestClamped = 3
 };
 
-inline constexpr u16 NUM_DEFAULT_SAMPLERS = 4;
+inline constexpr u32 NUM_DEFAULT_SAMPLERS = 4;
 
 struct Framebuffer
 {
@@ -375,9 +375,9 @@ struct GpuSystem
     gpu::Format::D16_UNORM_S8_UINT, gpu::Format::D24_UNORM_S8_UINT,
     gpu::Format::D32_SFLOAT_S8_UINT};
 
-  static constexpr u16 NUM_TEXTURE_SLOTS = 2'048;
+  static constexpr u32 NUM_TEXTURE_SLOTS = 2'048;
 
-  static constexpr u16 NUM_SAMPLER_SLOTS = 128;
+  static constexpr u32 NUM_SAMPLER_SLOTS = 128;
 
   static constexpr u16 NUM_FRAME_TIMESPANS = 2'048;
